+++ packages/universal-wallet-react/src/molecules/dialogs/ExploreDialog/ExploreDialog.stories.tsx
@@ -5,13 +5,13 @@ import { action } from '@storybook/addon-actions';
 import {
   FullscreenDialog as Component,
   Props,
-} from '../../atoms/FullscreenDialog';
-import { CapturePasswordDialogContent } from '../../atoms/CapturePasswordDialog/CapturePasswordDialogContent';
+} from '../../../atoms/FullscreenDialog';
+import { ExploreDialogContent } from './ExploreDialogContent';
 
-import { seedToId, passwordToKey } from '../../__fixtures__/wallet';
+import { fullWallet } from '../../../__fixtures__/wallet';
 
 const meta: Meta = {
-  title: 'Molecules/Dialogs/Export/Dialog',
+  title: 'Molecules/Dialogs/Explore/Dialog',
   component: Component,
 };
 
@@ -19,21 +19,15 @@ export default meta;
 
 const Template: Story<Props> = (args) => <Component {...args} />;
 
+// By passing using the Args format for exported stories, you can control the props for a component for reuse in a test
+// https://storybook.js.org/docs/react/workflows/unit-testing
 export const Dialog = Template.bind({});
 
 Dialog.args = {
   dialogTrigger: <Button>trigger</Button>,
-  dialogTitle: 'Export',
-  dialogSubmitTitle: 'Export',
-  dialogContent: (
-    <CapturePasswordDialogContent
-      seedToId={seedToId}
-      passwordToKey={passwordToKey}
-      setDialogState={(state: any) => {
-        action('setDialogState')(state);
-      }}
-    />
-  ),
+  dialogTitle: 'Explore',
+  dialogSubmitTitle: 'Explore',
+  dialogContent: <ExploreDialogContent contents={fullWallet.contents} />,
   handleSubmit: async () => {
     action('handleSubmit')();
   },
