+++ packages/universal-wallet-react/src/molecules/dialogs/ImportDialog/index.tsx
@@ -1,1 +1,2 @@
 export * from './ImportDialogContent';
+export * from './buildImportInterface';
